refactor (registration-api): Move Factories into model package
