+++ applications/registration-api/src/main/java/no/dcat/model/DatasetFactory.java
@@ -1,7 +1,5 @@
-package no.dcat.factory;
+package no.dcat.model;
 
-import no.dcat.model.Catalog;
-import no.dcat.model.Dataset;
 import org.springframework.beans.BeanUtils;
 
 import java.util.Date;
